requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.1
django-environ==0.12.0
django-markdownify==0.9.5
django-ninja==1.5.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
django-types==0.22.0
idna==3.11
Markdown==3.10
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
schema==0.7.8
sqlparse==0.5.3
tinycss2==1.4.0
types-psycopg2==2.9.21.20251012
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
